Refactor utils
Removed some `fs` calls that were not logically needed, split apart the utils into smalle r, separate pieces
